RELOCATE initialization part in other file
